Safari-609.1.11.1/LayoutTests/webarchive/loading/test-loading-archive-subresource-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
frame "<!--frame1-->" - didStartProvisionalLoadForFrame
main frame - didHandleOnloadEventsForFrame
frame "<!--frame1-->" - didFailProvisionalLoadWithError
main frame - didFinishLoadForFrame
 This test assumes the webarchive is not loaded.
